requirements.txt
pandas
python-binance
TA-Lib
numpy
os